+++ solvis_store/model.py
@@ -1,7 +1,7 @@
 import logging
 from datetime import datetime as dt
 
-from pynamodb.attributes import NumberAttribute, ListAttribute, NumberSetAttribute, UnicodeAttribute
+from pynamodb.attributes import ListAttribute, NumberAttribute, NumberSetAttribute, UnicodeAttribute
 from pynamodb.models import Model
 
 from .cloudwatch import ServerlessMetricWriter
@@ -34,7 +34,7 @@ class RuptureSetLocationDistances(MetricatedModel):
     radius = NumberAttribute()
     location = UnicodeAttribute()
     ruptures = NumberSetAttribute()  # Rupture Index,
-    distances = ListAttribute(of = NumberAttribute)  # optional list of distances, one for each rupture_index
+    distances = ListAttribute(of=NumberAttribute)  # distances, one for each rupture_index
     rupture_count = NumberAttribute()